Prototyoe version 1
Adnan, I have iterated the prototype and all the content is ready to be linked with each other in order to get a working a prototype. Kindly check it out and link the pages. If you can do it over the weekend that will be great. This way we will have some time to make changes in it before demo
Thanks